requirements.txt
alembic==1.10.3
aniso8601==9.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.6
PyJWT==2.6.0
pytest==7.3.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
requests-mock==1.10.0
six==1.16.0
SQLAlchemy==2.0.8
swagger-gen==0.1.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
waitress==2.1.2
Werkzeug==2.2.3
win32-setctime==1.1.0
